+++ types/valid-keys.ts
@@ -1,0 +1,8 @@
+export enum VALID_KEYS {
+  OPENAI_API_KEY = "OPENAI_API_KEY",
+  ANTHROPIC_API_KEY = "ANTHROPIC_API_KEY",
+  GOOGLE_GEMINI_API_KEY = "GOOGLE_GEMINI_API_KEY",
+  MISTRAL_API_KEY = "MISTRAL_API_KEY",
+  PERPLEXITY_API_KEY = "PERPLEXITY_API_KEY",
+  AZURE_OPENAI_API_KEY = "AZURE_OPENAI_API_KEY"
+}
